devfreq: Introduce devfreq boost driver
This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are
near-zero at the boosted frequencies and power consumption is minimized
at said frequencies. The goal of this driver is to provide an interface
to achieve optimal device performance by requesting boosts on key
events, such as when a frame is ready to rendered to the display.

Currently, support is only present for boosting the cpu-llcc-ddr-bw
devfreq device, but the driver is structured in a way that makes it
easy to add support for new boostable devfreq devices in the future.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

drm/atomic: Boost DDR bus when committing a new frame

In order to reduce jank, request DDR bus boosts whenever a new frame
is ready to be rendered to the display. The boosts should be sufficient
enough to render 60 FPS without any dropped frames when there is no
significant external source of load.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>

mm: Boost when memory pressure becomes high

When memory pressure starts to balloon, allocation latency takes a hit
as the page allocator continuously tries to find a way to free up some
pages. Boosting when this occurs can significantly help reduce
allocation latency, preventing noticeable lag from being observed.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>

Defconfig-v1 Enable devfreq boost DDR